+++ axelor-account/src/main/java/com/axelor/apps/account/web/InvoiceController.java
@@ -927,12 +927,16 @@ public class InvoiceController {
 
     long companyId = company.getPartner() == null ? 0 : company.getPartner().getId();
 
-    String domain =
-        String.format(
-            "self.id != %d AND self.isContact = false AND self.isCustomer = true", companyId);
+    String domain = String.format("self.id != %d AND self.isContact = false ", companyId);
     domain += " AND :company member of self.companySet";
 
     int invoiceTypeSelect = Beans.get(InvoiceService.class).getPurchaseTypeOrSaleType(invoice);
+
+    if (invoiceTypeSelect == 1) {
+      domain += " AND self.isCustomer = true ";
+    } else {
+      domain += " AND self.isSupplier = true ";
+    }
 
     try {