+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveServiceImpl.java
@@ -469,8 +469,9 @@ public class StockMoveServiceImpl implements StockMoveService {
     try {
       Beans.get(TemplateMessageService.class).generateAndSendMessage(stockMove, template);
     } catch (Exception e) {
-      throw new AxelorException(
-          TraceBackRepository.CATEGORY_CONFIGURATION_ERROR, e.getMessage(), stockMove);
+      //      throw new AxelorException(
+      //          TraceBackRepository.CATEGORY_CONFIGURATION_ERROR, e.getMessage(), stockMove);
+      LOG.error(e.getMessage());
     }
   }
 
@@ -627,7 +628,8 @@ public class StockMoveServiceImpl implements StockMoveService {
 
     stockMoveLines = MoreObjects.firstNonNull(stockMoveLines, Collections.emptyList());
     StockMove newStockMove = stockMoveRepo.copy(stockMove, false);
-
+    // In copy OriginTypeSelect set null.
+    newStockMove.setOriginTypeSelect(stockMove.getOriginTypeSelect());
     for (StockMoveLine stockMoveLine : stockMoveLines) {
 
       if (stockMoveLine.getQty().compareTo(stockMoveLine.getRealQty()) > 0) {