+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderInvoiceServiceImpl.java
@@ -46,6 +46,7 @@ import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderComputeService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderLineService;
+import com.axelor.apps.sale.service.saleorder.SaleOrderWorkflowServiceImpl;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.repo.StockMoveRepository;
 import com.axelor.apps.supplychain.db.Timetable;
@@ -92,6 +93,8 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
 
   protected StockMoveRepository stockMoveRepository;
 
+  protected SaleOrderWorkflowServiceImpl saleOrderWorkflowServiceImpl;
+
   @Inject
   public SaleOrderInvoiceServiceImpl(
       AppSupplychainService appSupplychainService,
@@ -99,15 +102,16 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
       InvoiceRepository invoiceRepo,
       InvoiceService invoiceService,
       SaleOrderLineService saleOrderLineService,
-      StockMoveRepository stockMoveRepository) {
+      StockMoveRepository stockMoveRepository,
+      SaleOrderWorkflowServiceImpl saleOrderWorkflowServiceImpl) {
 
     this.appSupplychainService = appSupplychainService;
-
     this.saleOrderRepo = saleOrderRepo;
     this.invoiceRepo = invoiceRepo;
     this.invoiceService = invoiceService;
     this.stockMoveRepository = stockMoveRepository;
     this.saleOrderLineService = saleOrderLineService;
+    this.saleOrderWorkflowServiceImpl = saleOrderWorkflowServiceImpl;
   }
 
   @Override
@@ -297,9 +301,13 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
 
     Invoice invoice = invoiceGenerator.generate();
 
-    invoiceGenerator.populate(
-        invoice,
-        this.createInvoiceLinesFromTax(invoice, taxLineList, invoicingProduct, percentToInvoice));
+    List<InvoiceLine> invoiceLinesList =
+        (taxLineList != null && !taxLineList.isEmpty())
+            ? this.createInvoiceLinesFromTax(
+                invoice, taxLineList, invoicingProduct, percentToInvoice)
+            : this.createInvoiceLinesFromSO(invoice, saleOrder, invoicingProduct, percentToInvoice);
+
+    invoiceGenerator.populate(invoice, invoiceLinesList);
 
     invoice.setAddressStr(saleOrder.getMainInvoicingAddressStr());
 
@@ -382,6 +390,54 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
       }
     }
     return createdInvoiceLineList;
+  }
+
+  protected List<InvoiceLine> createInvoiceLinesFromSO(
+      Invoice invoice, SaleOrder saleOrder, Product invoicingProduct, BigDecimal percentToInvoice)
+      throws AxelorException {
+
+    List<InvoiceLine> invoiceLineList = new ArrayList<>();
+    BigDecimal lineAmountToInvoice =
+        percentToInvoice
+            .multiply(saleOrder.getInTaxTotal())
+            .divide(new BigDecimal("100"), 4, BigDecimal.ROUND_HALF_EVEN);
+
+    InvoiceLineGenerator invoiceLineGenerator =
+        new InvoiceLineGenerator(
+            invoice,
+            invoicingProduct,
+            invoicingProduct.getName(),
+            lineAmountToInvoice,
+            lineAmountToInvoice,
+            lineAmountToInvoice,
+            invoicingProduct.getDescription(),
+            BigDecimal.ONE,
+            invoicingProduct.getUnit(),
+            null,
+            InvoiceLineGenerator.DEFAULT_SEQUENCE,
+            BigDecimal.ZERO,
+            PriceListLineRepository.AMOUNT_TYPE_NONE,
+            lineAmountToInvoice,
+            null,
+            false,
+            false,
+            0) {
+          @Override
+          public List<InvoiceLine> creates() throws AxelorException {
+
+            InvoiceLine invoiceLine = this.createInvoiceLine();
+
+            List<InvoiceLine> invoiceLines = new ArrayList<>();
+            invoiceLines.add(invoiceLine);
+
+            return invoiceLines;
+          }
+        };
+
+    List<InvoiceLine> invoiceOneLineList = invoiceLineGenerator.creates();
+    invoiceLineList.addAll(invoiceOneLineList);
+
+    return invoiceLineList;
   }
 
   @Override
@@ -674,6 +730,11 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
           saleOrder.getSaleOrderSeq());
     }
     saleOrder.setAmountInvoiced(amountInvoiced);
+
+    if (appSupplychainService.getAppSupplychain().getCompleteSaleOrderOnInvoicing()
+        && amountInvoiced.compareTo(saleOrder.getExTaxTotal()) == 0) {
+      saleOrderWorkflowServiceImpl.completeSaleOrder(saleOrder);
+    }
   }
 
   @Override